Initial Unit Test setup
Unfortunately this doesn't cover any Play-mode tests, as the requriement
for Playmode tests to use an Assembly definition complicates our setup a
lot.

We need to review SC as a whole and establish what we want to test and
on what level to determine if that's a thing we need to investigate
further. For now, this is sufficient for testing what we have.